Consume `Property` items from local blockchain, render data accordingly on Home page and Details page Closes #12
Now that the Leasy contract exposes the `getProperties` method, it can be consumed by the Frontend. This set of changes consists of calling the `getProperties` from the Leasy smart contract deployed at `NEXT_PUBLIC_LEASY_CONTRACT_ADDRESS` using the ABI in Foundry's output folder `contracts/out/` and rendering the `Property` items data accordingly on home page and the details page.

- create the `PropertiesContext`, implement the `PropertiesContextProvider` that uses `useContractRead` to call `getProperties` and returns a provider that exposes the fetches `Property` items to children components
- wrap the root component in `layout.tsx` with the `PropertiesContextProvider` so that `Property` items are available to any components that may needed it
- call `usePropertiesContext()` on the details page to look up the property identified by the `propertyID` instead of using the fixtures array
- call `usePropertiesContext()` on the home page to list the properties
- added `localhost` chain to the wagmi chains config.