Implement the `signLease` contract method Closes #19
## Descripton
Once the leasing process was initiated for a property via the `leaseProperty` contract method call, decisions/signatures can then be submitted. This set of changes add the `signLease` `Leasy` contract method
that takes a property ID and a signature status as input arguments that persists a decision for the sender for a property for which the leasing process was initiated.

## Changes in this Pull Request
- some refactoring to make it easier to maintain primarily by introducing some new fields
- validate the property exists
- validate the leasing process was initiated for the property
- validate the sender is one of the designated renters for the property
- update the signature status of the renter for the property