Render property status in property card and on property details page, render property signers and signature statuses on details page Closes #16
## Description
A property owner may initiate the leasing process by calling the new `leaseProperty` method of the `Leasy` contract.
`Property`s now have a status, a renters array and a signature statuses array.
This set of changes consists of rendering a property status in the property card and on property details page, as well as renderering the property signers and signature statuses on details page.

## Changes in this Pull Request
- modify the `Property` model to reflect the field name changes in the contract
- add the `status` and `depositAmount` to the `Property` model
- add the `PropertyStatusBadge` component that renders a colored badge based a property status
- render the property status badge in the property card
- render the property status badge on the property details page
- render the list of renters and their corresponding signature statuses on the property details page