Add ability to users to apply for a property for specific dates and for owners to designate applicants as renters Closes #23, #24
## Description
Currently, the application is more about signing lease agreements than renting properties. This set of changes applying a relatively minor set of changes to transition to a peer-to-peer renting model instead. Specifically, a contract method is added allowing users to inquire for a stay at a property for a specific set of dates and a property owner to designate applicants as renters for specific dates. The concept of "signatures" and "lease agreement" is being completed removed.

## Changes in this Pull Request
- add the `inquireStay` contract method that allows a user to add themself as an applicant for a property for specific dates if the property exists and that it is not inactive (that it is available)
- refactored the `assignRenters` method by renaming it to `selectApplicant` and changing the logic to transfer applicants (new concept) to renters
- removed all the logic around signatures and lease agreements