Require property deposit to be submitted with application (#28)
Closes #27

## Description
There is now a contract method allowing applicants to send an inquiry for a stay at a property for a specific set of dates. This task consists of requiring each applicant to submit the deposit amount with their application. Submitting an application with less than the deposit amount defined for a property should revert with an error.

## Changes in this Pull Request
- add the `InsufficientDeposit` that throws when a call is made to `inquireStay` with an Ether value that is less than the property's required deposit amount
- keep track of deposits submitted by property by applicants in new `applicantsDeposits` field
- make `inquireStay` `payable`, revert if `msg.value` is less than the property's required deposit amount. If valid, then keep track of value in new `applicantsDeposits` field