Support navigation to property details page directly via URL (#31)
* Support navigation to property details page directly via URL

Closes #30

## Description
Users can currently navigate to a property details page only from the dashboard.

This set of changes consists of consuming the path parameter and to get the property by ID on page load so that users can also directly navigate to a property details page via URL (eg. `/property/1`).

## Changes in this Pull Request
- add `getProperty` Leasy contract method to get property by ID
- modify the `PropertyDetails` component to fetch the property by ID from the contract instead of getting it from the properties context

* Support navigation to property details page directly via URL

Closes #30

## Description
Users can currently navigate to a property details page only from the dashboard.

This set of changes consists of consuming the path parameter and to get the property by ID on page load so that users can also directly navigate to a property details page via URL (eg. `/property/1`).

## Changes in this Pull Request
- add `getProperty` Leasy contract method to get property by ID
- modify the `PropertyDetails` component to fetch the property by ID from the contract instead of getting it from the properties context